% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bounds.R
\name{filter_bounds}
\alias{filter_bounds}
\title{Filter image collection using bounds of a provided region (feature)}
\usage{
filter_bounds(image_collection, region)
}
\description{
Filter image collection using bounds of a provided region (feature)
}
\author{
Alec L. Robitaille
}
