% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date.R
\name{filter_date}
\alias{filter_date}
\title{Filter image collection within start and end date}
\usage{
filter_date(image_collection, start_date, end_date)
}
\description{
Filter image collection within start and end date
}
\author{
Alec L. Robitaille
}
