% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ee_image_collection.R
\name{get_ee_image_collection}
\alias{get_ee_image_collection}
\title{Get EE image collection}
\usage{
get_ee_image_collection(asset_id)
}
\description{
Get EE image collection
}
\author{
Alec L. Robitaille
}
