% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_features.R
\name{get_example_features}
\alias{get_example_features}
\title{Get example features}
\usage{
get_example_features(x, y, type = "points", region_size, n_features)
}
\arguments{
\item{type}{either 'points', or 'polygons'}

\item{region_size}{size of example study region's (in this case, a circle's diameter in meters)}
}
\description{
Get example features
}
\author{
Alec L. Robitaille
}
