% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bands_image_collection.R
\name{select_bands_image_collection}
\alias{select_bands_image_collection}
\title{Select bands from  image collection}
\usage{
select_bands_image_collection(image_collection, band_names)
}
\description{
Select bands from  image collection
}
\author{
Alec L. Robitaille
}
